requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
openpyxl
xlrd
PyMuPDF
SQLAlchemy